file view changes

+++ server/controllers/fileController.js
@@ -111,6 +111,7 @@ const getFiles = asyncHandler(async (req, res) => {
 		],
 	})
 		.select('-originalName')
+		.sort({ createdAt: -1 })
 		.lean()
 		.exec();
 
@@ -217,7 +218,8 @@ const deleteFile = asyncHandler(async (req, res) => {
 		return res.status(400).json({ message: 'File not found' });
 	}
 
-	await file.deleteOne().exec();
+	if (userId === file.createdBy) await file.deleteOne().exec();
+	else await file.updateOne({ sharedWith: null }).exec();
 
 	await saveActivity({
 		userId: userId,
